Niloofar Sadeghi / Refactor who-we-are page (#2729)
* refactor who-we-are page

* destructuring maps

* add data types

* fix keys issue

* fix leaders avatar sizes in android
