+++ src/pages/who-we-are/_MakeTrading.tsx
@@ -11,7 +11,7 @@ const FirstSectionContainer = styled(SectionContainer)`
     flex-direction: column;
     align-items: center;
     padding: 120px 16px 24px;
-    @media ${device.tablet} {
+    @media ${device.tabletL} {
         padding: 40px 16px;
     }
 `
@@ -38,9 +38,9 @@ const StyledFirstSectionText = styled(Header)`
 const MakeTrading = () => {
     return (
         <FirstSectionContainer padding="120px 0 80px" background="var(--color-white)">
-            {first_section_texts.map((item, index) => (
+            {first_section_texts.map(({ text }, index) => (
                 <StyledFirstSectionText as="p" type="unset" key={index} size="24px">
-                    {item.text}
+                    {text}
                 </StyledFirstSectionText>
             ))}
             <StyledHeader as="h2" size="48px" align="center" type="page-title">
