+++ src/pages/who-we-are/_data.tsx
@@ -1,11 +1,16 @@
 import React, { ReactElement } from 'react'
-import { Localize } from 'components/localization'
+import { Localize, localize } from 'components/localization'
 import Shield from 'images/svg/who-we-are/shield.svg'
 import Star from 'images/svg/who-we-are/star.svg'
 import People from 'images/svg/who-we-are/people.svg'
 import Hands from 'images/svg/who-we-are/hands.svg'
+import { all_offices, all_countries } from 'pages/careers/_model/_locations/_locations'
 
-export const first_section_texts = [
+type FirstSectionTextsType = {
+    text: ReactElement
+}
+
+export const first_section_texts: FirstSectionTextsType[] = [
     {
         text: (
             <Localize translate_text="Deriv is one of the world’s largest online brokers. We offer CFDs and other derivatives on forex, indices, cryptocurrencies, commodities, and synthetics to millions of registered users across the globe." />
@@ -252,4 +257,25 @@ export const mobile_pins: MapPinType[] = [
         title: <Localize translate_text="Guernsey" />,
         link: '/careers/locations/guernsey',
     },
+]
+
+type DerivNumbersType = {
+    count: string
+    title: string
+}
+
+export const our_offices_count: DerivNumbersType[] = [
+    { count: localize('750+'), title: localize('employees') },
+    { count: localize('50+'), title: localize('nationalities') },
+    { count: localize(all_offices.length), title: localize('locations') },
+    { count: localize(all_countries.length), title: localize('countries') },
+]
+
+export const deriv_numbers: DerivNumbersType[][] = [
+    [
+        { count: localize('2.5M+'), title: localize('traders worldwide') },
+        { count: localize('USD 26M+'), title: localize('withdrawals last month') },
+        { count: localize('114M+'), title: localize('trades last month') },
+        { count: localize('USD 10B+'), title: localize('total trade turnover') },
+    ],
 ]
