+++ src/pages/who-we-are/_lazy-load.ts
@@ -1,0 +1,8 @@
+import Loadable from '@loadable/component'
+
+export const OurValues = Loadable(() => import('./_OurValues'))
+export const OurPrinciples = Loadable(() => import('./_OurPrinciples'))
+export const OurLeadership = Loadable(() => import('./_OurLeadership'))
+export const AboutUsBanner = Loadable(() => import('./_AboutUsBanner'))
+export const DerivNumbers = Loadable(() => import('./_DerivNumbers'))
+export const OurOffices = Loadable(() => import('./_OurOffices'))
