+++ src/pages/who-we-are/index.tsx
@@ -1,27 +1,27 @@
 import React from 'react'
 import styled from 'styled-components'
-import Loadable from '@loadable/component'
 import MakeTrading from './_MakeTrading'
 import Hero from './components/_hero'
 import ImageMarquee from './carousel/_ImageMarquee'
+import {
+    OurValues,
+    OurPrinciples,
+    OurLeadership,
+    AboutUsBanner,
+    DerivNumbers,
+    OurOffices,
+} from './_lazy-load'
 import device from 'themes/device'
 import { SEO } from 'components/containers'
 import Layout from 'components/layout/layout'
 import { localize, WithIntl } from 'components/localization'
-
-const OurValues = Loadable(() => import('./_OurValues'))
-const OurPrinciples = Loadable(() => import('./_OurPrinciples'))
-const OurLeadership = Loadable(() => import('./_OurLeadership'))
-const AboutUsBanner = Loadable(() => import('./_AboutUsBanner'))
-const DerivNumbers = Loadable(() => import('./_DerivNumbers'))
-const OurOffices = Loadable(() => import('./_OurOffices'))
 
 const StartSeparator = styled.div`
     width: 0;
     height: 0;
     border-top: 120px solid transparent;
     border-right: 99vw solid var(--color-grey-30);
-    @media ${device.tablet} {
+    @media ${device.tabletL} {
         border-top: 40px solid transparent;
     }
 `
@@ -31,7 +31,7 @@ const EndSeparator = styled.div`
     height: 0;
     border-left: 99vw solid var(--color-grey-30);
     border-bottom: 120px solid transparent;
-    @media ${device.tablet} {
+    @media ${device.tabletL} {
         border-bottom: 40px solid transparent;
     }
 `
